+++ Assets/AnimationClipToVrma.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 516ca51a19fbb8045b8efcce0ae89429
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
