+++ Assets/AnimationClipToVrma/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 492e6b2b0f5b1884a98d893e8c674aed
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
