+++ Assets/AnimationClipToVrma/Scripts/Editor/Util/AvatarDescriptionDetailLogger.cs
